chore: set USER in Dockerfile.konflux
We starting setting `runAsNonRoot: true` in the model registry
deployment (per RHOAIENG-13857). Unfortunately, that broke the model
registry operator tests because they use Kind, which starts containers
as `root` by default if neither the image nor the pod spec specifies the
user to run. We tried setting the user ID in the spec, but then we had
errors because the allowed UID ranges vary between Open Shift clusters.
Long story short, it would be helpful to put a `USER` directive in the
Dockerfile so that Kind and OpenShift will behave consistently.

Signed-off-by: Paul Boyd <pboyd@redhat.com>